rules: mirror the DerivedLevel test into rule_test.go
